BG1-87 - Billing System: Show discount percentage and allow to change it

+++ particular-service/src/main/java/com/rr/particularservice/model/Particular.java
@@ -22,6 +22,6 @@ public class Particular {
     private String particularName;
 
     @Column(name = "DiscountPercentage")
-    private int discountPercentage;
+    private float discountPercentage;
 
 }
